requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
colorama==0.4.4
discord-py-interactions==4.0.2
discord.py==1.7.3
google-api-core==2.5.0
google-api-python-client==2.38.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
multidict==5.2.0
oauthlib==3.2.0
orjson==3.6.3
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
wincertstore==0.2
yarl==1.7.2
